+++ .gitignore
@@ -1,22 +1,13 @@
-.env
-# dependencies
-/node_modules
-/.pnp
-.pnp.js
+# next.js
+.next/
+out/
 
-# testing
-/coverage
+# Modules
+node_modules/
 
-# production
-/build
-
-# misc
-.DS_Store
+# Local environment variables
 .env.local
-.env.development.local
-.env.test.local
-.env.production.local
 
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
+# System files
+.DS_Store
+*.log
