+++ components/Footer.tsx
@@ -1,0 +1,333 @@
+import React, { useState } from 'react';
+import Link from 'next/link';
+
+const Footer: React.FC = () => {
+    const iconSize = 'h-8 w-8';
+    const socialMediaIconClass = `text-white p-2 rounded-full flex items-center justify-center ${iconSize} transition-transform transform hover:scale-110`;
+    const buttonClass = `
+    px-6 py-4 
+    text-sm font-semibold 
+    text-gray-100 
+    bg-purple-gray 
+    rounded-md 
+    hover:bg-purple-gray 
+    opacity-80 hover:opacity-100 
+    w-full md:w-auto
+    disabled:opacity-50 disabled:cursor-not-allowed
+  `.trim();
+    const footerInputClass = `
+    px-4 py-4 
+    text-sm text-gray-900 
+    placeholder-gray-400 
+    bg-gray-100 
+    border border-gray-300 
+    rounded-md 
+    dark:text-gray-400 
+    dark:placeholder-gray-400 
+    dark:bg-gray-700 
+    dark:border-gray-700 
+    w-full md:w-auto
+    md:w-96
+  `.trim();
+
+    const [email, setEmail] = useState('');
+    const [isSubmitting, setIsSubmitting] = useState(false);
+    const [submitStatus, setSubmitStatus] = useState<'idle' | 'success' | 'error'>('idle');
+    const [error, setError] = useState<string | null>(null);
+
+    const handleEmailChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+        setEmail(event.target.value);
+        if (submitStatus !== 'idle') {
+            setSubmitStatus('idle');
+        }
+        if (error) {
+            setError(null);
+        }
+    };
+
+    const validateEmail = (email: string) => {
+        const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+        return emailRegex.test(email);
+    };
+
+    const handleSubmit = async () => {
+        const trimmedEmail = email.trim();
+
+        if (!trimmedEmail) {
+            setError("Please provide your email to join us.");
+            setSubmitStatus('error');
+            return;
+        }
+
+        if (!validateEmail(trimmedEmail)) {
+            setError("Please enter a valid email address.");
+            setSubmitStatus('error');
+            return;
+        }
+
+        setIsSubmitting(true);
+        setError(null);
+        setSubmitStatus('idle');
+
+        try {
+            // Use environment variables with Next.js naming convention
+            const API_URL = process.env.NEXT_PUBLIC_API_URL;
+            const API_KEY = process.env.NEXT_PUBLIC_API_KEY;
+
+            // Check if API configuration is available
+            if (!API_URL || !API_KEY) {
+                console.log("API configuration not set, using fallback method");
+                throw new Error("API configuration is missing or not set");
+            }
+
+            const response = await fetch(API_URL, {
+                method: "POST",
+                headers: {
+                    Authorization: `Bearer ${API_KEY}`,
+                    "Content-Type": "application/json",
+                    Accept: "application/json",
+                },
+                body: JSON.stringify({
+                    email: trimmedEmail,
+                    groups: [],
+                    trigger_automation: true,
+                }),
+            });
+
+            if (response.ok) {
+                console.log("Successfully submitted user data to API");
+                setSubmitStatus('success');
+                setEmail('');
+                setError(null);
+            } else {
+                const errorData = await response.json();
+                console.error("API error:", errorData);
+                setError("There was an error submitting your email. Please try again.");
+                setSubmitStatus('error');
+            }
+        } catch (err) {
+            console.error("Submission error:", err);
+
+            // Fallback to the original method if API fails
+            try {
+                const response = await fetch('https://stats.sender.net/forms/aKrmkz/subscribe', {
+                    method: 'POST',
+                    headers: {
+                        'Content-Type': 'application/json',
+                    },
+                    body: JSON.stringify({
+                        email: trimmedEmail
+                    })
+                });
+
+                if (response.ok) {
+                    setSubmitStatus('success');
+                    setEmail('');
+                    setError(null);
+                } else {
+                    setError("There was an error submitting your email. Please try again.");
+                    setSubmitStatus('error');
+                }
+            } catch (fallbackError) {
+                console.error('Fallback subscription error:', fallbackError);
+                setError("There was an error submitting your email. Please try again.");
+                setSubmitStatus('error');
+            }
+        } finally {
+            setIsSubmitting(false);
+        }
+    };
+
+    const handleKeyPress = (event: React.KeyboardEvent<HTMLInputElement>) => {
+        if (event.key === 'Enter') {
+            handleSubmit();
+        }
+    };
+
+    const getButtonText = () => {
+        if (isSubmitting) return 'JOINING...';
+        if (submitStatus === 'success') return 'JOINED!';
+        if (submitStatus === 'error') return 'TRY AGAIN';
+        return 'JOIN US';
+    };
+
+    // Social media icons as SVG components
+    const FacebookIcon = () => (
+        <svg className="w-4 h-4" fill="currentColor" viewBox="0 0 24 24">
+            <path d="M24 12.073c0-6.627-5.373-12-12-12s-12 5.373-12 12c0 5.99 4.388 10.954 10.125 11.854v-8.385H7.078v-3.47h3.047V9.43c0-3.007 1.792-4.669 4.533-4.669 1.312 0 2.686.235 2.686.235v2.953H15.83c-1.491 0-1.956.925-1.956 1.874v2.25h3.328l-.532 3.47h-2.796v8.385C19.612 23.027 24 18.062 24 12.073z" />
+        </svg>
+    );
+
+    const InstagramIcon = () => (
+        <svg className="w-4 h-4" fill="currentColor" viewBox="0 0 24 24">
+            <path d="M12 2.163c3.204 0 3.584.012 4.85.07 3.252.148 4.771 1.691 4.919 4.919.058 1.265.069 1.645.069 4.849 0 3.205-.012 3.584-.069 4.849-.149 3.225-1.664 4.771-4.919 4.919-1.266.058-1.644.07-4.85.07-3.204 0-3.584-.012-4.849-.07-3.26-.149-4.771-1.699-4.919-4.92-.058-1.265-.07-1.644-.07-4.849 0-3.204.013-3.583.07-4.849.149-3.227 1.664-4.771 4.919-4.919 1.266-.057 1.645-.069 4.849-.069zM12 0C8.741 0 8.333.014 7.053.072 2.695.272.273 2.69.073 7.052.014 8.333 0 8.741 0 12c0 3.259.014 3.668.072 4.948.2 4.358 2.618 6.78 6.98 6.98C8.333 23.986 8.741 24 12 24c3.259 0 3.668-.014 4.948-.072 4.354-.2 6.782-2.618 6.979-6.98.059-1.28.073-1.689.073-4.948 0-3.259-.014-3.667-.072-4.947-.196-4.354-2.617-6.78-6.979-6.98C15.668.014 15.259 0 12 0zm0 5.838a6.162 6.162 0 100 12.324 6.162 6.162 0 000-12.324zM12 16a4 4 0 110-8 4 4 0 010 8zm6.406-11.845a1.44 1.44 0 100 2.881 1.44 1.44 0 000-2.881z" />
+        </svg>
+    );
+
+    const LinkedInIcon = () => (
+        <svg className="w-4 h-4" fill="currentColor" viewBox="0 0 24 24">
+            <path d="M20.447 20.452h-3.554v-5.569c0-1.328-.027-3.037-1.852-3.037-1.853 0-2.136 1.445-2.136 2.939v5.667H9.351V9h3.414v1.561h.046c.477-.9 1.637-1.85 3.37-1.85 3.601 0 4.267 2.37 4.267 5.455v6.286zM5.337 7.433a2.062 2.062 0 01-2.063-2.065 2.064 2.064 0 112.063 2.065zm1.782 13.019H3.555V9h3.564v11.452zM22.225 0H1.771C.792 0 0 .774 0 1.729v20.542C0 23.227.792 24 1.771 24h20.451C23.2 24 24 23.227 24 22.271V1.729C24 .774 23.2 0 22.222 0h.003z" />
+        </svg>
+    );
+
+    const YouTubeIcon = () => (
+        <svg className="w-4 h-4" fill="currentColor" viewBox="0 0 24 24">
+            <path d="M23.498 6.186a3.016 3.016 0 0 0-2.122-2.136C19.505 3.545 12 3.545 12 3.545s-7.505 0-9.377.505A3.017 3.017 0 0 0 .502 6.186C0 8.07 0 12 0 12s0 3.93.502 5.814a3.016 3.016 0 0 0 2.122 2.136c1.871.505 9.376.505 9.376.505s7.505 0 9.377-.505a3.015 3.015 0 0 0 2.122-2.136C24 15.93 24 12 24 12s0-3.93-.502-5.814zM9.545 15.568V8.432L15.818 12l-6.273 3.568z" />
+        </svg>
+    );
+
+    return (
+        <footer className="bg-gray-800 text-white">
+            <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+                <section className="py-10 sm:pt-16 lg:pt-24">
+                    <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-y-16 lg:gap-x-12 text-center">
+                        <div className="flex flex-col items-center h-full space-y-6">
+                            <p className="text-sm font-semibold tracking-widest text-gray-400 uppercase">
+                                Pivot For Dancers
+                            </p>
+                            <p className="text-gray-300 leading-relaxed">
+                                Discover inspiration, guidance, and a vibrant community to help you navigate your journey toward a fulfilling and meaningful future filled with purpose, growth, opportunities, and lasting connections.
+                            </p>
+                        </div>
+
+                        <div className="flex flex-col items-center space-y-6 lg:px-6">
+                            <p className="text-sm font-semibold tracking-widest text-gray-400 uppercase">
+                                Join our Community
+                            </p>
+                            <div className="flex flex-col space-y-4 w-full max-w-sm">
+                                <input
+                                    type="email"
+                                    placeholder="Enter your email"
+                                    value={email}
+                                    onChange={handleEmailChange}
+                                    onKeyPress={handleKeyPress}
+                                    disabled={isSubmitting}
+                                    className={`${footerInputClass} h-12`}
+                                />
+                                <button
+                                    type="button"
+                                    onClick={handleSubmit}
+                                    disabled={isSubmitting || !email.trim()}
+                                    className={`${buttonClass} h-12 flex items-center justify-center`}
+                                >
+                                    {getButtonText()}
+                                </button>
+                            </div>
+
+                            {submitStatus === 'success' && (
+                                <p className="text-purple-gray text-sm font-medium">
+                                    Thank you! You should receive a confirmation shortly.
+                                </p>
+                            )}
+                            {error && (
+                                <p className="text-red-400 text-sm font-medium">
+                                    {error}
+                                </p>
+                            )}
+                        </div>
+
+                        <div className="flex flex-col items-center space-y-6">
+                            <p className="text-sm font-semibold tracking-widest text-gray-400 uppercase">
+                                Contact
+                            </p>
+                            <a
+                                href="mailto:pivotfordancers@gmail.com"
+                                className="text-gray-300 hover:text-white transition-colors"
+                            >
+                                pivotfordancers@gmail.com
+                            </a>
+                            <a
+                                href="https://instagram.com/pivotfordancers"
+                                target="_blank"
+                                rel="noopener noreferrer"
+                                className="text-gray-300 hover:text-white transition-colors"
+                            >
+                                @pivotfordancers
+                            </a>
+                            <ul className="flex items-center justify-center space-x-3">
+                                <li>
+                                    <a
+                                        href="https://www.facebook.com/pivotfordancers/"
+                                        className={`${socialMediaIconClass} bg-blue-500 hover:bg-blue-400`}
+                                        target="_blank"
+                                        rel="noopener noreferrer"
+                                        aria-label="Visit us on Facebook"
+                                    >
+                                        <FacebookIcon />
+                                    </a>
+                                </li>
+                                <li>
+                                    <a
+                                        href="https://www.instagram.com/pivotfordancers/"
+                                        className={`${socialMediaIconClass} bg-pink-500 hover:bg-pink-400`}
+                                        target="_blank"
+                                        rel="noopener noreferrer"
+                                        aria-label="Visit us on Instagram"
+                                    >
+                                        <InstagramIcon />
+                                    </a>
+                                </li>
+                                <li>
+                                    <a
+                                        href="https://www.linkedin.com/company/pivotfordancers/"
+                                        className={`${socialMediaIconClass} bg-blue-600 hover:bg-blue-500`}
+                                        target="_blank"
+                                        rel="noopener noreferrer"
+                                        aria-label="Visit us on LinkedIn"
+                                    >
+                                        <LinkedInIcon />
+                                    </a>
+                                </li>
+                                <li>
+                                    <a
+                                        href="https://www.youtube.com/@pivotfordancers"
+                                        className={`${socialMediaIconClass} bg-red-600 hover:bg-red-500`}
+                                        target="_blank"
+                                        rel="noopener noreferrer"
+                                        aria-label="Visit us on YouTube"
+                                    >
+                                        <YouTubeIcon />
+                                    </a>
+                                </li>
+                            </ul>
+                        </div>
+                    </div>
+                </section>
+
+                <div className="border-t border-gray-600 my-1"></div>
+
+                <div className="py-12 text-center space-y-4">
+                    <div className="flex flex-col md:flex-row justify-center items-center gap-6">
+                        <a
+                            href="https://drive.google.com/file/d/1XiqWl5qiHd1xCGF62D300hyodmsCe5OC/view?usp=sharing"
+                            target="_blank"
+                            rel="noopener noreferrer"
+                            className="text-gray-400 hover:text-gray-300 transition"
+                        >
+                            Privacy Policy
+                        </a>
+                        <a
+                            href="https://drive.google.com/file/d/1JVBQN7SZBSSup9E-avhsTXRxZ1Y_7Zeg/view?usp=sharing"
+                            target="_blank"
+                            rel="noopener noreferrer"
+                            className="text-gray-400 hover:text-gray-300 transition"
+                        >
+                            Happy Trails Terms and Conditions
+                        </a>
+                        <a
+                            href="https://drive.google.com/file/d/11ssn_alwWa1K0IrgGQgE7Dt65TRmsXrH/view?usp=sharing"
+                            target="_blank"
+                            rel="noopener noreferrer"
+                            className="text-gray-400 hover:text-gray-300 transition"
+                        >
+                            Pivot Mentorship Program Terms and Conditions
+                        </a>
+                    </div>
+                    <p className="text-sm text-gray-400">&copy; {new Date().getFullYear()} Pivot for Dancers. All rights reserved.</p>
+                </div>
+            </div>
+        </footer>
+    );
+};
+
+export default Footer;
