+++ package.json
@@ -19,9 +19,10 @@
     "framer-motion": "^12.7.4",
     "gh-pages": "^6.3.0",
     "lucide-react": "^0.501.0",
+    "next": "^15.5.3",
     "postcss": "^8.4.31",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-icons": "^5.0.1",
     "react-intersection-observer": "^9.5.3",
     "react-scripts": "5.0.1",
@@ -36,12 +37,10 @@
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "CI='' react-scripts build",
-    "predeploy": "npm run build",
-    "deploy": "gh-pages -d build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
   },
   "eslintConfig": {
     "extends": [
