--- src/react-app-env.d.ts
@@ -1,7 +1,0 @@
-/// <reference types="react-scripts" />
-declare namespace NodeJS {
-    interface ProcessEnv {
-        REACT_APP_API_URL: string;
-        REACT_APP_API_KEY: string;
-    }
-}
